% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean raw earthquake data}
\usage{
eq_clean_data(df_raw)
}
\arguments{
\item{df_raw}{dataframe of raw NOAA data}
}
\value{
Returns a cleaned version of df_raw
}
\description{
Cleans raw data by creating a date column with year, month and day in one date, converting
LATITUDE and LONGITUDE columns to numeric and cleaning the LOCATION_NAME column with the eq_location_clean() function.
Negative years have been removed.
}
\examples{
\dontrun{
  df <- load_data() \%>\% dplyr::filter(YEAR >= 0) \%>\%
       dplyr::mutate(date = get_date(DAY, MONTH, YEAR))
  print(class(df))
  head(df)
}

}
