% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Generate earthquake label column}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{dataframe for creating annotation column}
}
\value{
Returns a vector of earthquake details
}
\description{
Input a dataframe (df) and the function will output a vector with HTML column names
Required columns: CLEAN_LOCATION_NAME, EQ_PRIMARY and DEATHS
}
\examples{
\dontrun{
map2 <-load_data() \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "GREECE" & lubridate::year(date) >= 2002) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
print(map2)
}
}
