% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean LOCATION_NAME column}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{dataframe containing LOCATION_NAME column}
}
\value{
Returns df with cleaned LOCATION_NAME column
}
\description{
This function cleans the LOCATION_NAME column by removing country name and
converting names to title case.
}
\examples{
\dontrun{
  df <- load_data() \%>\% eq_location_clean()
  print(class(df))
  head(df)
}

}
