% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_maps.R
\name{eq_map}
\alias{eq_map}
\title{Generate leaflet map with annotations}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{dataframe for plotting}

\item{annot_col}{column from input dataframe for annotation}
}
\value{
Returns an interactive leaflet map
}
\description{
Subset of earthquake data plotted on a map using leaflet.
Input dataframe requires LONGITUDE, LATITUDE and EQ_PRIMARY columns.
Other required argument: annotation column ('annot_col')
}
\examples{
\dontrun{
 map1 <-load_data() \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "GREECE" & lubridate::year(date) >= 2002) \%>\%
   eq_map(annot_col = "date")
 print(map1)
}
}
