% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timelines.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label function}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = 5,
  ...
)
}
\arguments{
\item{n_max}{Integer (default = 5); represents the max number of earthquakes to show.}
}
\description{
Run geom_timeline first to create timeline.
This geom will add a vertical line and label to those points.
}
\examples{
\dontrun{
ggplot(df, aes(x = date, y = COUNTRY,
               color = as.numeric(TOTAL_DEATHS),
               size = as.numeric(EQ_PRIMARY),
               label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  labs(size = "Richter scale value", color = "# deaths") +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
        legend.position = "bottom",
        axis.title.y = ggplot2::element_blank()) + ggplot2::xlab("DATE") +
  geom_timeline_label(data=df)
}
}
