% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timelines.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Define GeomTimeline class}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
The GeomTimeline charts a timeline of earthquakes for a given country
with points along the timeline for each earthquake. The size of the point inicates
the magnitude of the earthquake, while the color indicates the number of deaths.
The date (x) is a required aesthetic. The country (y) is optional.
}
\examples{
\dontrun{
 df <- df_earthquakes \%>\% filter(COUNTRY \%in\% c("GREECE", "USA"), YEAR > 2002)
 ggplot(df, aes(x = date, y = COUNTRY,
                color = as.numeric(TOTAL_DEATHS),
                size = as.numeric(EQ_PRIMARY),
                label = CLEAN_LOCATION_NAME)) +
   geom_timeline() +
   labs(size = "Richter scale value", color = "# deaths") +
   ggplot2::theme(panel.background = ggplot2::element_blank(),
         legend.position = "bottom",
         axis.title.y = ggplot2::element_blank()) +
   ggplot2::xlab("DATE")
}
}
\keyword{datasets}
