% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{get_date}
\alias{get_date}
\title{Generate a single date column from NOAA data's DAY, MONTH and YEAR columns}
\usage{
get_date(days, months, years)
}
\arguments{
\item{days}{day of year as (int)}

\item{months}{month (int)}

\item{years}{year (int)}
}
\value{
Returns a single date vector DATES
}
\description{
Takes the three date-related columns and turns them into one cohesive date column. Some instances of
DAY and MONTH are NA. NA days will be replaced with 1, NA months will be replaced with Jan.
Negative years have been removed.
}
\examples{
\dontrun{
  df <- load_data() \%>\% dplyr::filter(YEAR >= 0) \%>\%
       dplyr::mutate(date = get_date(DAY, MONTH, YEAR))
  print(class(df))
  head(df)
}

}
