% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{load_data}
\alias{load_data}
\title{Load raw data}
\usage{
load_data(path = file.path("inst/extdata", "signif.txt.tsv"))
}
\arguments{
\item{path}{file path to data source (default == "inst/extdata/signif.txt.tsv")}
}
\value{
Returns a dataframe
}
\description{
Function loads raw, tab delimited data (from NOAA, in data folder).
}
\examples{
\dontrun{
  df <- load_data()
  print(class(df))
  head(df)
}

}
