% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timelines.R
\name{plot_earthquakes_timeline_label}
\alias{plot_earthquakes_timeline_label}
\title{Plot  earthquake timeline with labels}
\usage{
plot_earthquakes_timeline_label(df, save_png = FALSE)
}
\arguments{
\item{df}{Earthquake data}

\item{save_png}{Indicates whether to save plot}
}
\description{
This plots and formats the earthquake data, with the addition of labels for the points along the timeline.
The plot will be saved as earthquakes_timeline_label.png if save_png = TRUE.
}
\examples{
\dontrun{
df <- df_earthquakes \%>\%
  filter(COUNTRY \%in\% c("GREECE", "USA"), YEAR > 2002)
plot_earthquakes_timeline_label(df, save_png=TRUE)
}
}
